🏭 Refactor: Reorganize components of AccApproval
Separate components that require rerendering from components that do not, to eliminate unnecessary rerendering.
